Adding secondary level status code for saml response. - renaming InvalidSamlOperationException to UnsuccessfulSamlOperationException - moving exception classes to a separate folder Exceptions, also tests are moved, changing namespaces - renaming Saml2Response_GetClaims_ThrowsOnStatusFailure2() to Saml2Response_GetClaims_ThrowsOnStatusFailure2()
