+++ Kentor.AuthServices.Tests/Kentor.AuthServices.Tests.csproj
@@ -116,13 +116,13 @@
     <Compile Include="HttpModule\HttpRequestBaseExtensionsTests.cs" />
     <Compile Include="Internal\QueryStringHelperTests.cs" />
     <Compile Include="Internal\PathHelperTests.cs" />
-    <Compile Include="InvalidSamlOperationExceptionTest.cs" />
+    <Compile Include="Exceptions\UnsuccessfulSamlOperationExceptionTest.cs" />
     <Compile Include="Metadata\MetadatabaseExtensionsTests.cs" />
     <Compile Include="Mvc\AuthServicesControllerTests.cs" />
     <Compile Include="ClaimsAuthenticationManagerStub.cs" />
     <Compile Include="ExceptionTestHelpers.cs" />
     <Compile Include="Saml2P\FilteringXmlNodeReaderTests.cs" />
-    <Compile Include="Saml2ResponseFailedValidationExceptionTests.cs" />
+    <Compile Include="Exceptions\Saml2ResponseFailedValidationExceptionTests.cs" />
     <Compile Include="WebSSO\CommandResultTests.cs" />
     <Compile Include="WebSSO\MetadataCommandTests.cs" />
     <Compile Include="ExtendedMetadataSerializerTests.cs" />
@@ -143,10 +143,10 @@
     <Compile Include="WebSSO\Saml2BindingTests.cs" />
     <Compile Include="ClaimsIdentityExtensionsTests.cs" />
     <Compile Include="Saml2AssertionExtensionsTests.cs" />
-    <Compile Include="AuthServicesExceptionTests.cs" />
-    <Compile Include="BadFormatSamlResponseExceptionTests.cs" />
+    <Compile Include="Exceptions\AuthServicesExceptionTests.cs" />
+    <Compile Include="Exceptions\BadFormatSamlResponseExceptionTests.cs" />
     <Compile Include="Internal\EnumeratorTests.cs" />
-    <Compile Include="NoSamlResponseFoundExceptionTests.cs" />
+    <Compile Include="Exceptions\NoSamlResponseFoundExceptionTests.cs" />
     <Compile Include="Internal\ReturnRequestedIssuerNameRegistryTests.cs" />
     <Compile Include="WebSSO\Saml2RedirectBindingTests.cs" />
     <Compile Include="AllTestsOtherCultures.cs" />
