+++ Kentor.AuthServices/Exceptions/BadFormatSamlResponseException.cs
@@ -5,25 +5,25 @@ using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Kentor.AuthServices
+namespace Kentor.AuthServices.Exceptions
 {
     /// <summary>
-    /// A SAML2 Response failed validation.
+    /// A SAML response was found, but could not be parsed due to formatting issues.
     /// </summary>
     [Serializable]
-    public class Saml2ResponseFailedValidationException : AuthServicesException
+    public class BadFormatSamlResponseException: AuthServicesException
     {
         /// <summary>
         /// Ctor
         /// </summary>
-        public Saml2ResponseFailedValidationException()
+        public BadFormatSamlResponseException()
         { }
 
         /// <summary>
         /// Ctor
         /// </summary>
         /// <param name="message">Message of the exception.</param>
-        public Saml2ResponseFailedValidationException(string message) : base(message)
+        public BadFormatSamlResponseException(string message) : base(message)
         { }
 
         /// <summary>
@@ -31,8 +31,8 @@ namespace Kentor.AuthServices
         /// </summary>
         /// <param name="message">Message of the exception.</param>
         /// <param name="innerException">Inner exception.</param>
-        public Saml2ResponseFailedValidationException(string message, Exception innerException)
-            : base(message, innerException)
+        public BadFormatSamlResponseException(string message, Exception innerException)
+            :base(message, innerException)
         { }
 
         /// <summary>
@@ -40,7 +40,7 @@ namespace Kentor.AuthServices
         /// </summary>
         /// <param name="info">Serialization info</param>
         /// <param name="context">Serialization context</param>
-        protected Saml2ResponseFailedValidationException(SerializationInfo info, StreamingContext context)
+        protected BadFormatSamlResponseException(SerializationInfo info, StreamingContext context)
             : base(info, context)
         { }
     }
